+++ daskms/__init__.py
@@ -1,10 +1,16 @@
 # -*- coding: utf-8 -*-
 
+# Activate the casacure backend (DASK_MS_BACKEND=casacure) before any
+# module pulls in casacore.tables via lazy_import.
+from daskms.casacure_backend import activate_casacure_backend
+
 import logging
+
+activate_casacure_backend()
 
 __author__ = """Simon Perkins"""
 __email__ = "sperkins@ska.ac.za"
-__version__ = "0.2.23"
+__version__ = "0.2.32"
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())